+++ src/main/java/project/volunteer/domain/notice/api/dto/response/NoticeListResponse.java
@@ -1,16 +1,15 @@
-package project.volunteer.domain.notice.api.dto;
+package project.volunteer.domain.notice.api.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import project.volunteer.domain.notice.application.dto.NoticeDetails;
 
+import java.util.List;
+
+@Getter
 @AllArgsConstructor
 @NoArgsConstructor
-@Getter
-public class NoticeDetailsResponse {
-
-    NoticeDetails notice;
-    //댓글 details dto 추가 필요
-
+public class NoticeListResponse {
+    List<NoticeDetails> noticeList;
 }
