+++ src/main/java/project/volunteer/domain/notice/application/NoticeService.java
@@ -1,7 +1,7 @@
 package project.volunteer.domain.notice.application;
 
-import project.volunteer.domain.notice.api.dto.NoticeAdd;
-import project.volunteer.domain.notice.api.dto.NoticeEdit;
+import project.volunteer.domain.notice.api.dto.request.NoticeAdd;
+import project.volunteer.domain.notice.api.dto.request.NoticeEdit;
 import project.volunteer.domain.notice.domain.Notice;
 
 public interface NoticeService {
