+++ src/main/java/project/volunteer/domain/signup/api/dto/response/KakaoUserInfoResponse.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import project.volunteer.domain.signup.dto.KakaoUserInfo;
-import project.volunteer.global.common.response.BaseResponse;
+import project.volunteer.global.common.dto.BaseResponse;
 
 @Getter
 @Setter
