+++ src/main/java/project/volunteer/domain/signup/api/dto/response/UserSaveResponse.java
@@ -3,7 +3,7 @@ package project.volunteer.domain.signup.api.dto.response;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import project.volunteer.global.common.response.BaseResponse;
+import project.volunteer.global.common.dto.BaseResponse;
 
 @Getter 
 @Setter
