+++ src/main/java/project/volunteer/global/common/dto/CommentContentParam.java
@@ -1,4 +1,4 @@
-package project.volunteer.domain.logboard.api.dto.request;
+package project.volunteer.global.common.dto;
 
 import javax.validation.constraints.NotNull;
