+++ src/main/java/project/volunteer/global/jwt/api/dto/AccessTokenReissueResponse.java
@@ -3,7 +3,7 @@ package project.volunteer.global.jwt.api.dto;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import project.volunteer.global.common.response.BaseResponse;
+import project.volunteer.global.common.dto.BaseResponse;
 
 @Getter @Setter
 @NoArgsConstructor
