+++ src/test/java/project/volunteer/domain/notice/api/NoticeControllerTest.java
@@ -17,8 +17,8 @@ import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.transaction.annotation.Transactional;
 import project.volunteer.domain.confirmation.dao.ConfirmationRepository;
 import project.volunteer.domain.confirmation.domain.Confirmation;
-import project.volunteer.domain.notice.api.dto.NoticeAdd;
-import project.volunteer.domain.notice.api.dto.NoticeEdit;
+import project.volunteer.domain.notice.api.dto.request.NoticeAdd;
+import project.volunteer.domain.notice.api.dto.request.NoticeEdit;
 import project.volunteer.domain.notice.dao.NoticeRepository;
 import project.volunteer.domain.notice.domain.Notice;
 import project.volunteer.domain.recruitment.dao.RecruitmentRepository;
