+++ src/test/java/project/volunteer/domain/sehedule/application/ScheduleDtoServiceImplTest.java
@@ -24,7 +24,7 @@ import project.volunteer.domain.user.domain.Gender;
 import project.volunteer.domain.user.domain.Role;
 import project.volunteer.domain.user.domain.User;
 import project.volunteer.global.common.component.*;
-import project.volunteer.global.common.response.StateResponse;
+import project.volunteer.global.common.dto.StateResponse;
 import project.volunteer.global.error.exception.BusinessException;
 
 import javax.persistence.EntityManager;
